Split different algos in different source files
